% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra algorithm}
\source{
See more at: \url{https://en.wikipedia.org/wiki/Dijkstras_algorithm}
}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph, }{data.frame with three variables (v1, v2 and w) that contains the edges of the graph (from
v1 to v2) with the weight of the edge (w).}

\item{init_node, }{numeric scalar.}
}
\value{
shortest path from the
initial node to every other node in the graph.
}
\description{
\code{dijkstra} algorithm takes a graph and an initial node and calculates the shortest path from the
initial node to every other node in the graph
}
\details{
The algorithm exists in many variants. Dijkstra's original algorithm found 
the shortest path between two given nodes, but a more common variant fixes 
a single node as the "source" node and finds shortest paths from the source to 
all other nodes in the graph, producing a shortest-path tree.
}
\examples{
wiki_graph <-
data.frame(v1=c(1,1,1,2,2,2,3,3,3,3,4,4,4,5,5,6,6,6),
          v2=c(2,3,6,1,3,4,1,2,4,6,2,3,5,4,6,1,3,5),
          w=c(7,9,14,7,10,15,9,10,11,2,15,11,6,6,9,14,2,9))

dijkstra(wiki_graph, 1)
dijkstra(wiki_graph, 3)


}
