% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab3elife599carde734-package.R
\docType{package}
\name{lab3elife599carde734-package}
\alias{lab3elife599carde734}
\alias{lab3elife599carde734-package}
\title{lab3elife599carde734: Package for lab3 containing functions euclidean and dijkstra}
\description{
Package for lab3 containing functions euclidean and dijkstra
}
\details{
Package for lab3 containing functions euclidean and dijkstra
}
\author{
\strong{Maintainer}: Elisio Quintino \email{elife599@student.liu.se}

Authors:
\itemize{
  \item Carolina Rosario \email{carde734@student.liu.se}
}

}
\keyword{internal}
