% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab3elife599carde734-package.R
\docType{package}
\name{lab3elife599carde734-package}
\alias{lab3elife599carde734}
\alias{lab3elife599carde734-package}
\title{lab3elife599carde734: Package for assignment 3 in Advanced R}
\description{
Package for assignment 3 in Advanced R. It contains functions euclidean and dijkstra, as well as dataset wiki_graph.
}
\details{
This package contains functions and dataset for Assignment 3 in Advanced R.
It contains two functions, euclidean and dijkstra, as well as the dataset wiki_graph.
Read more about them in the respective documentations.
}
\author{
\strong{Maintainer}: Elisio Quintino \email{elife599@student.liu.se}

Authors:
\itemize{
  \item Carolina Rosario \email{carde734@student.liu.se}
}

}
\keyword{internal}
