\name{lab3elife599carde734-package}
\alias{lab3elife599carde734-package}
\alias{lab3elife599carde734}
\docType{package}
\title{
\packageTitle{lab3elife599carde734}
}
\description{
\packageDescription{lab3elife599carde734}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lab3elife599carde734}
\packageIndices{lab3elife599carde734}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{lab3elife599carde734}

Maintainer: \packageMaintainer{lab3elife599carde734}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
